+++ packages/desktop_autopaste/lib/desktop_autopaste.dart
@@ -21,6 +21,17 @@ class DesktopAutopaste {
     );
   }
 
+  Future<bool> pasteFromClipboard({
+    Duration prePasteDelay = Duration.zero,
+    DesktopAutopastePasteShortcut pasteShortcut =
+        DesktopAutopastePasteShortcut.ctrlV,
+  }) {
+    return DesktopAutopasteFfi.instance.pasteFromClipboard(
+      prePasteDelay: prePasteDelay,
+      pasteShortcut: pasteShortcut,
+    );
+  }
+
   Future<FocusedTextFieldContext> getFocusedTextFieldContext({
     int? maxCharsBefore = 120,
     int? maxCharsAfter = 120,